failed attempt to udpate dockerfile
This is extremely frustrating at the moement. I am getting a connection
reset by peer error and I do not know why. For some reason on the main
styles container, I cannot get the docker container to run even after
I've cleaned out all of my docker containers. I keep getting the same
Liquid exception that I've seen before and I'm sure I fixed, but it's
not clear at all and I know that I'm in the flailing stage of trying to
debug something, but I'm not there yet and I'm not sure where I can go
to debug all of this.