don't require password
This ultimately assumes that the users are siloed on their own machines.
In the future, it would be good to add an option for them to turn this
on or off, but for the moment, we are setting this to avoid confusion
for users who are not terribly comfortable with the command line.